feat: fakes
- Centralize the fake value generation to a fakes file for book.
- Update the various places in the code to use this new fake
